Add additional check for DNS propagation
Before checking the dns visability from a list of random servers the propagation
of the challenge dns entries must be checked against the authoritative nameservers.

If the public dns visability is checked before the dns entry is propagated to the
responisible nameservers the caching dns servers queries the domain and will
receive a NXDOMAIN. This will lead to a negative caching. The negative cache
has it's own TTL defined in the SOA record. This is normaly between 900 and 3600 seconds
or even longer. If the same caching nameserver is queried a second time
it serves the NXDOMAIN information for at least the negative TTL.
To avoid this issue wait for the propagation.
Upper timeout limit is 240 seconds. It soft failes if the propagation timed out. The dns
visability check will hard fail as a backup.